repo updates

+++ UserServiceDAL/Repositories/UserRepository.cs
@@ -9,16 +9,17 @@ public class UserRepository : IUserRepository
 {
     private readonly AppDbContext _context;
     public UserRepository(AppDbContext context) => _context = context;
-
     public async Task<IEnumerable<User>> GetAllAsync() => await _context.Users.ToListAsync();
-    public async Task<User> GetByIdAsync(Guid id) => await _context.Users.FindAsync(id);
-    public async Task<User> GetByEmailAsync(string email) => await _context.Users
+    public async Task<User?> GetByIdAsync(Guid id) => await _context.Users.FindAsync(id) ?? null;
+    public async Task<User?> GetByEmailAsync(string email) => await _context.Users
         .FirstOrDefaultAsync(u => u.Email == email);
     public async Task AddAsync(User user) => await _context.Users.AddAsync(user);
-    public async Task UpdateAsync(User user) => _context.Users.Update(user);
+    public void Update(User user) => _context.Users.Update(user);
     public async Task DeleteAsync(Guid id)
     {
         var user = await GetByIdAsync(id);
+        if (user == null)
+            throw new Exception();
         _context.Users.Remove(user);
     }
 }
